Move shared/game/util.ts to renderer.
