+++ src/renderer/util/game.ts
